+++ arguments/arguments.go
@@ -31,6 +31,7 @@ type ParserArguments struct {
 	LegacyPush              bool
 	NoRoute                 bool
 	AddRoutes               bool
+	NoStart                 bool
 }
 
 type stringSlice []string
@@ -77,6 +78,7 @@ func ParseArgs(args []string) (*ParserArguments, error) {
 	flags.BoolVar(&pta.LegacyPush, "legacy-push", false, "use legacy push instead of new v3 api")
 	flags.BoolVar(&pta.NoRoute, "no-routes", false, "deploy new application without adding routes")
 	flags.BoolVar(&pta.AddRoutes, "add-routes", false, "add routes from manifest to app only")
+	flags.BoolVar(&pta.NoStart, "no-start", false, "don't start application after deployment")
 	//flags.BoolVar(&pta.ShowLogs, "show-app-log", false, "tail and show application log during application start")
 	//flags.StringVar(&pta.DockerImage, "docker-image", "", "url to docker image")
 	//flags.StringVar(&pta.DockerUserName, "docker-username", "", "pass docker username if image came from private repository")